requirements.txt
Django
django-cors-headers
django-ninja
django-ninja-jwt
pydantic[email]
psycopg2-binary
psycopg
Pillow
rav